backend/TakeHome/requirements.txt
Django==2.2
mysqlclient==1.4.6
django-environ==0.4.5
Pillow==7.1.2
pytesseract==0.3.4
django-pure-pagination==0.3.0
